refactor: move listCommands/listCommandNames into CommandRegistry
Resolves the follow-up from #618 review: `listCommandNames` and
`listCommands` are now methods on `CommandRegistry` instead of
standalone functions in `input.ts` that took the registry as their
first argument. This makes `CommandRegistry` the single place to
query all available commands.

Along with the methods, the related helpers (`parseFrontmatter`,
`resolveCommandFilePath`, `resolveContent`, `listSkillNames`,
`ListDir`, `CommandSuggestion`) are moved to `command-registry.ts`
and re-exported from `input.ts` for backward compatibility. The
injectable `listDir`/`readFile` params are preserved as method
arguments so testability is unchanged.

Closes #621
